+++ src/parseTools.js
@@ -1037,7 +1037,7 @@ Please update to new syntax.`);
     if (DYNCALLS) {
       return `(function(cb, ${args}) { ${returnExpr} getDynCaller("${sig}", cb)(${args}) })`;
     } else {
-      return `(function(cb, ${args}) { ${returnExpr} wasmTable.get(cb)(${args}) })`;
+      return `(function(cb, ${args}) { ${returnExpr} wbind(cb)(${args}) })`;
     }
   }
 
@@ -1049,7 +1049,7 @@ Please update to new syntax.`);
       return `(function() { ${returnExpr} ${dyncall}.call(null, ${funcPtr}); })`;
     }
   } else {
-    return `wasmTable.get(${funcPtr})`;
+    return `wbind(${funcPtr})`;
   }
 }